# server/internal/controller/transport/http/openapi/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Project Template API",
		"version": "0.1.0",
		"description": "Contract for the Project Template modular monolith."
	},
...
	"paths": {
...
		"/workspaces": {
			"get": {
				"operationId": "listWorkspaces",
				"summary": "List workspaces",
				"parameters": [],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"post": {
				"operationId": "createWorkspace",
				"summary": "Create workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"201": {
						"description": "The request has succeeded and a new resource has been created as a result.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWorkspaceRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
		"/workspaces/{workspaceId}": {
			"get": {
				"operationId": "getWorkspace",
				"summary": "Get workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceResponse"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"patch": {
				"operationId": "updateWorkspace",
				"summary": "Update workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"200": {
						"description": "The request has succeeded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkspaceResponse"
								}
							}
						}
					},
					"400": {
						"description": "The server could not understand the request due to invalid syntax.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkspaceRequest"
							}
						}
					}
				},
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			},
			"delete": {
				"operationId": "deleteWorkspace",
				"summary": "Delete workspace",
				"parameters": [
					{
						"$ref": "#/components/parameters/WorkspaceIdPath"
					},
					{
						"$ref": "#/components/parameters/CSRFHeader"
					}
				],
				"responses": {
					"204": {
						"description": "There is no content to send for this request, but the headers may be useful. "
					},
					"401": {
						"description": "Access is unauthorized.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"403": {
						"description": "Access is forbidden.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "The server cannot find the requested resource.",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"422": {
						"description": "Client error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				},
				"tags": ["Workspaces"],
				"security": [
					{
						"SessionCookieAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"parameters": {
			"CSRFHeader": {
				"name": "X-CSRF-Token",
				"in": "header",
				"required": true,
				"schema": {
					"type": "string",
					"minLength": 1
				}
			},
...
			"WorkspaceIdPath": {
				"name": "workspaceId",
				"in": "path",
				"required": true,
				"schema": {
					"$ref": "#/components/schemas/uuid"
				}
			}
		},
		"schemas": {
...
			"CreateWorkspaceRequest": {
				"type": "object",
				"required": ["name"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 80
					}
				},
				"examples": [
					{
						"name": "Product Engineering"
					}
				]
			},
...
			"ProblemDetails": {
				"type": "object",
				"required": ["type", "title", "status", "code"],
				"properties": {
					"type": {
						"type": "string",
						"format": "uri"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer",
						"format": "int32"
					},
					"code": {
						"type": "string",
						"enum": [
							"BAD_REQUEST",
							"UNAUTHORIZED",
							"FORBIDDEN",
							"NOT_FOUND",
							"CONFLICT",
							"VALIDATION_FAILED",
							"INTERNAL_ERROR",
							"SERVICE_UNAVAILABLE",
							"METHOD_NOT_ALLOWED",
							"AUTH_MISSING_SESSION",
							"AUTH_INVALID_SESSION",
							"AUTH_INVALID_CREDENTIALS",
							"AUTH_INVALID_CSRF",
							"INSUFFICIENT_ROLE",
							"EMAIL_ALREADY_EXISTS",
							"WORKSPACE_NOT_FOUND",
							"WORKSPACE_MEMBER_NOT_FOUND",
							"WORKSPACE_MEMBER_ALREADY_EXISTS",
							"WORKSPACE_LAST_OWNER",
							"TASK_ASSIGNEE_NOT_MEMBER",
							"TASK_NOT_FOUND"
						]
					},
					"detail": {
						"type": "string"
					},
					"requestId": {
						"type": "string"
					},
					"errors": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProblemError"
						}
					}
				},
				"examples": [
					{
						"type": "https://project-template.example/problems/validation-failed",
						"title": "Validation failed",
						"status": 422,
						"code": "VALIDATION_FAILED",
						"detail": "One or more request fields are invalid.",
						"requestId": "req_01J00000000000000000000000",
						"errors": [
							{
								"code": "REQUIRED",
								"message": "title is required",
								"location": "body.title"
							}
						]
					}
				]
			},
			"ProblemError": {
				"type": "object",
				"required": ["code", "message"],
				"properties": {
					"code": {
						"type": "string",
						"enum": [
							"INVALID_VALUE",
							"INVALID_JSON",
							"INVALID_LENGTH",
							"INVALID_FORMAT",
							"INVALID_ENUM",
							"NOT_A_WORKSPACE_MEMBER",
							"REQUIRED",
							"VALUE_TOO_SHORT",
							"VALUE_TOO_LONG"
						]
					},
					"message": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"examples": [
					{
						"code": "REQUIRED",
						"message": "title is required",
						"location": "body.title"
					}
				]
			},
...
			"UpdateWorkspaceRequest": {
				"type": "object",
				"required": ["name"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 80
					}
				},
				"examples": [
					{
						"name": "Platform Engineering"
					}
				]
			},
...
			"Workspace": {
				"type": "object",
				"required": ["id", "name", "role", "createdByUserId", "createdAt", "updatedAt"],
				"properties": {
					"id": {
						"$ref": "#/components/schemas/uuid"
					},
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 80
					},
					"role": {
						"type": "string",
						"enum": ["owner", "editor", "viewer"]
					},
					"createdByUserId": {
						"$ref": "#/components/schemas/uuid"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"examples": [
					{
						"id": "22222222-2222-2222-2222-222222222222",
						"name": "Product Engineering",
						"role": "owner",
						"createdByUserId": "11111111-1111-1111-1111-111111111111",
						"createdAt": "2026-07-10T08:10:00Z",
						"updatedAt": "2026-07-10T08:10:00Z"
					}
				]
			},
			"WorkspaceListResponse": {
				"type": "object",
				"required": ["workspaces"],
				"properties": {
					"workspaces": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Workspace"
						}
					}
				}
			},
...
			"WorkspaceResponse": {
				"type": "object",
				"required": ["workspace"],
				"properties": {
					"workspace": {
						"$ref": "#/components/schemas/Workspace"
					}
				}
			},
...
			"uuid": {
				"type": "string",
				"format": "uuid"
			}
		},
		"securitySchemes": {
			"SessionCookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "__Host-project_template_session"
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1",
			"description": "Version 1 API",
			"variables": {}
		}
	]
}